refactor: Update fridge feature and remove outdated files
