--- app/src/main/res/layout/activity_fridge.xml
@@ -1,180 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#FFFFFF"
-    android:padding="16dp"
-    tools:context=".fridge.FridgeActivity">
-
-    <!-- 뒤로가기 아이콘 -->
-    <ImageView
-        android:id="@+id/backArrow"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="28dp"
-        android:src="@drawable/ic_arrow_back"
-        android:contentDescription="Back Arrow"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <!-- 상단 제목 (Fridge) -->
-    <TextView
-        android:id="@+id/textViewTitle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="100dp"
-        android:layout_marginTop="16dp"
-        android:text="Fridge"
-        android:textColor="#575757"
-        android:textSize="32sp"
-        android:textStyle="bold"
-        app:layout_constraintStart_toEndOf="@+id/backArrow"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <!-- 검색/입력 칸 (길게 하나) + 오른쪽에 (⊕) + (돋보기) -->
-    <LinearLayout
-        android:id="@+id/linearSearchContainer"
-        android:layout_width="match_parent"
-        android:layout_height="48dp"
-        android:layout_marginTop="35dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginEnd="8dp"
-        android:background="@drawable/rounded_border_white"
-        android:orientation="horizontal"
-        android:paddingStart="12dp"
-        android:paddingEnd="8dp"
-        android:gravity="center_vertical"
-        app:layout_constraintTop_toBottomOf="@+id/textViewTitle"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent">
-
-        <!-- 검색/입력 (EditText) - 최대 15자 -->
-        <EditText
-            android:id="@+id/editTextSearch"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:background="@android:color/transparent"
-            android:hint="재료 입력 (최대 15자)"
-            android:maxLength="15"
-            android:padding="0dp"
-            android:textColor="#000000"
-            android:textSize="15sp"
-            android:textColorHint="#9E9E9E"
-            android:importantForAutofill="noExcludeDescendants"
-            android:autofillHints="" />
-
-        <!-- 동그라미 플러스(⊕) 버튼 -->
-        <ImageView
-            android:id="@+id/buttonAddItem"
-            android:layout_width="40dp"
-            android:layout_height="40dp"
-            android:src="@drawable/ic_add"
-            android:contentDescription="Add Item"
-            android:layout_marginStart="4dp"
-            android:padding="8dp" />
-
-        <!-- 돋보기(검색) 버튼 -->
-        <ImageView
-            android:id="@+id/buttonSearch"
-            android:layout_width="40dp"
-            android:layout_height="40dp"
-            android:src="@drawable/ic_search"
-            android:contentDescription="Search"
-            android:layout_marginStart="4dp"
-            android:padding="8dp" />
-    </LinearLayout>
-
-    <!-- 스크롤 영역(재료 목록) -->
-    <ScrollView
-        android:id="@+id/scrollViewIngredients"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginTop="16dp"
-        android:fadeScrollbars="false"
-        android:scrollbarStyle="outsideOverlay"
-        app:layout_constraintTop_toBottomOf="@+id/linearSearchContainer"
-        app:layout_constraintBottom_toTopOf="@+id/bottomDivider"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent">
-
-        <!-- 동적으로 추가할 LinearLayout -->
-        <LinearLayout
-            android:id="@+id/linearLayoutList"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical" />
-
-    </ScrollView>
-
-    <!-- 하단 구분선(얇은 회색 선) -->
-    <View
-        android:id="@+id/bottomDivider"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:background="#D0D0D0"
-        app:layout_constraintBottom_toTopOf="@+id/bottomNavigation"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <!-- 하단 네비게이션 바(4개의 아이콘) -->
-    <LinearLayout
-        android:id="@+id/bottomNavigation"
-        android:layout_width="match_parent"
-        android:layout_height="60dp"
-        android:orientation="horizontal"
-        android:gravity="center"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent">
-
-        <!-- 홈 아이콘 -->
-        <ImageView
-            android:id="@+id/navHome"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:src="@drawable/ic_home"
-            android:contentDescription="Home"
-            android:padding="8dp"
-            android:scaleType="centerInside" />
-
-        <!-- 레시피 요청 아이콘 -->
-        <ImageView
-            android:id="@+id/navRecipeRequest"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:src="@drawable/ic_recipe_request"
-            android:contentDescription="Recipe Request"
-            android:padding="3dp"
-            android:scaleType="centerInside" />
-
-        <!-- 냉장고 아이콘 (현재 페이지) -->
-        <ImageView
-            android:id="@+id/navFridge"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:src="@drawable/ic_fridge"
-            android:contentDescription="Fridge"
-            android:padding="8dp"
-            android:scaleType="centerInside" />
-
-        <!-- 레시피 저장 아이콘 -->
-        <ImageView
-            android:id="@+id/navSavedRecipes"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:src="@drawable/ic_saved_recipe"
-            android:contentDescription="Saved Recipes"
-            android:padding="8dp"
-            android:scaleType="centerInside" />
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
